Update body of script in index.html
We create a var 'promise' and set the the output/resolve to this var. We handle the success/resolve of 'promise' by using a .then() and 
creating a blob (an object of raw data). This also returns a promise. We handle this promise, again,  using .then() so that we can 
process the URL of the image, create a <img> element and set the src attribute to the URL.